services/indexar_movies/requirements.txt
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.1
setuptools==70.1.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
